initial ColorRamp
